@@ -1,6 +1,8 @@
 # frozen_string_literal: true
+
 class UserRequestsController < ApplicationController
-  include ApplicationModule, UserModule
+  include ApplicationModule
+  include UserModule
 
   before_action :user_logged_in?
   before_action :check_if_user_on_team, only: [:create]
@@ -20,11 +22,11 @@ class UserRequestsController < ApplicationController
   # Allows the captain of a team to accept a request to join a team.
   def accept
     if @user_request.user_on_team?
-      redirect_to :back, alert: I18n.t('requests.accepted_another')
+      redirect_back(fallback_location: user_root_path, alert: I18n.t('requests.accepted_another'))
     elsif @user_request.accept
-      redirect_to team_path(@user_request.team), notice: I18n.t('requests.accepted_successful')
+      redirect_to user_root_path, notice: I18n.t('requests.accepted_successful')
     else
-      redirect_to :back, alert: I18n.t('requests.too_many_players')
+      redirect_back(fallback_location: user_root_path, alert: I18n.t('requests.too_many_players'))
     end
   end
 
@@ -32,7 +34,7 @@ class UserRequestsController < ApplicationController
   def destroy
     @user_request.status = :Rejected
     if @user_request.save
-      redirect_to :back, notice: I18n.t('requests.rejected_successful')
+      redirect_back(fallback_location: user_root_path, notice: I18n.t('requests.rejected_successful'))
     else
       redirect_to user_root_path, alert: @user_request.errors.full_messages.first
     end
@@ -50,8 +52,7 @@ class UserRequestsController < ApplicationController
   # Only allow team captain and the user requesting to delete a user request.
   def check_if_able_to_reject
     @user_request = UserRequest.find(params[:id])
-    unless @user_request.user.eql?(current_user) || @user_request.team.team_captain.eql?(current_user)
-      raise ActiveRecord::RecordNotFound
-    end
+    return true if @user_request.user.eql?(current_user) || @user_request.team.team_captain.eql?(current_user)
+    raise ActiveRecord::RecordNotFound
   end
 end
