@@ -1,10 +1,11 @@
 # frozen_string_literal: true
+
 # Every time a team captain invites a user to their team, a new user invite is created.
 class UserInvite < ActiveRecord::Base
   belongs_to :team
   belongs_to :user
 
-  enum status: [:Pending, :Accepted, :Rejected]
+  enum status: %i[Pending Accepted Rejected]
 
   after_create :link_to_user, :send_email
 
@@ -23,9 +24,8 @@ class UserInvite < ActiveRecord::Base
 
   # Make sure a user cannot be invited to the same team over and over.
   def uniqueness_of_invite
-    unless UserInvite.pending.where(team: team, email: email).empty? && team.users.where(email: email).empty?
-      errors.add(:email, :uniqueness)
-    end
+    return true if UserInvite.pending.where(team: team, email: email).empty? && team.users.where(email: email).empty?
+    errors.add(:email, :uniqueness)
   end
 
   def accept
@@ -34,7 +34,7 @@ class UserInvite < ActiveRecord::Base
     elsif user.nil? || !user.team.nil? # Check to make sure user isn't already on a team.
       false
     else
-      update_attribute(:status, :Accepted)
+      update_attributes(status: :Accepted)
       team.users << user
     end
   end
@@ -46,6 +46,6 @@ class UserInvite < ActiveRecord::Base
   end
 
   def link_to_user
-    update_attribute(:user, User.find_by_email(email))
+    update_attributes(user: User.find_by(email: email))
   end
 end
