@@ -1,7 +1,7 @@
 # frozen_string_literal: true
+
 # Controller for the homepage. Since there is not much logic on the homepage this controller is
 # rather empty.
 class HomeController < ApplicationController
-  def index
-  end
+  def index; end
 end
