@@ -1,10 +1,11 @@
 # frozen_string_literal: true
+
 # Whenever a user requests to join a team a new user request is created.
 class UserRequest < ActiveRecord::Base
   belongs_to :team
   belongs_to :user
 
-  enum status: [:Pending, :Accepted, :Rejected]
+  enum status: %i[Pending Accepted Rejected]
 
   after_create :send_email
 
@@ -21,9 +22,8 @@ class UserRequest < ActiveRecord::Base
 
   # Make sure a user cannot be invited to the same team over and over.
   def uniqueness_of_pending_request
-    unless UserRequest.pending_requests.where(team: team, user: user).empty?
-      errors.add(:user, I18n.t('requests.already_pending'))
-    end
+    return true if UserRequest.pending_requests.where(team: team, user: user).empty?
+    errors.add(:user, I18n.t('requests.already_pending'))
   end
 
   def user_on_team?
@@ -36,7 +36,7 @@ class UserRequest < ActiveRecord::Base
     elsif user_on_team? # Check to make sure user isn't already on a team.
       false
     else
-      update_attribute(:status, :Accepted)
+      update_attributes(status: :Accepted)
       team.users << user
     end
   end
